+++ HISTORY.rst
@@ -9,6 +9,7 @@ Changelog
 * If ``required=True`` validation is violated for a given ``Field``, it will raise an error message that is different from the message specified by the ``error`` argument. Thanks Anders Steinlein.
 * More generic error message raised when required field is missing.
 * ``validated`` decorator should only wrap a ``Field`` class's output method.
+* Add ``Serializer.error_handler`` decorator that allows user to specify a custom error handler.
 
 0.6.0 (2014-06-03)
 ++++++++++++++++++